requirements.txt
python-vlc
colorama
